requirements.txt
# MiniDriveWorld 依赖

# 核心框架
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
h5py>=3.9.0

# nuScenes 数据集
nuscenes-devkit>=1.1.10

# 模型相关
transformers>=4.30.0
diffusers>=0.20.0
timm>=0.9.0
einops>=0.6.0

# 训练工具
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.65.0

# 推理部署
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
# tensorrt  # 需要单独安装

# API 服务
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 工具
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# 开发工具
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0

# Jupyter
jupyter>=1.0.0
ipywidgets>=8.1.0
